% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{areColours}
\alias{areColours}
\title{Validate vector of colours}
\usage{
areColours(x)
}
\arguments{
\item{x}{A vector of objects that can be coerced as a colour by \code{col2rgb()} (see \code{?col2rgb} for more information)}
}
\value{
Named vector of logical values
}
\description{
Adapted from \href{https://stackoverflow.com/questions/13289009/check-if-character-string-is-a-valid-color-representation}{this StackOverflow post}
}
\examples{
my_colours <- c("Colour1" = "#FFFFFF", "Colour2" = "#2", "Colour3" = "snow1", "Colour4" = "2")

areColours(my_colours)
}
