% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uq.R
\name{colours_uq}
\alias{colours_uq}
\title{Retrieve UQ colour palettes, apply transparency}
\usage{
colours_uq(
  ...,
  palette = c("main", "secondary", "bw", "discrete", "continuous", "neutral"),
  alpha = 1
)
}
\arguments{
\item{...}{(Optional) Specify subset of palette colours to include}

\item{palette}{Palette name (one of \code{main}, \code{secondary}, \code{bw}, \code{discrete}, \code{continuous}, or \code{neutral}). Default: \code{main}}

\item{alpha}{Colour transparency. Default: 1 (no transparency)}
}
\value{
Named vector of colours
}
\description{
Retrieve UQ colour palettes, apply transparency
}
\examples{
library(ggplot2)

# No alpha, default palette (`main`)
ggplot(mtcars, aes(x = cyl, y = mean(qsec))) +
 geom_bar(stat = "identity",
          fill = colours_uq())

# You can change the transparency using `alpha`
ggplot(mtcars, aes(x = cyl, y = mean(qsec))) +
 geom_bar(stat = "identity",
          fill = colours_uq(alpha = 0.5))

# You can also use specific colours from a palette
ggplot(mtcars, aes(x = cyl, y = mean(qsec))) +
 geom_bar(stat = "identity",
          fill = colours_uq("Light purple", palette = "secondary"))
}
