% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uq.R
\name{scale_colour_uq_d}
\alias{scale_colour_uq_d}
\title{Discrete ggplot2 palette for \code{colour}}
\usage{
scale_colour_uq_d(
  palette = c("main", "secondary", "bw", "discrete", "continuous", "neutral"),
  alpha = 1,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Palette name (one of \code{main}, \code{secondary}, \code{bw}, \code{discrete}, \code{continuous}, or \code{neutral}). Default: \code{main}}

\item{alpha}{Colour transparency. Default: 1 (no transparency)}

\item{reverse}{If \code{TRUE}, reverse the colour palette order. Default: \code{FALSE}}

\item{...}{Additional arguments to pass to \code{ggplot2::discrete_scale()}}
}
\value{
A discrete ggplot2 palette for \code{colour}
}
\description{
Discrete ggplot2 palette for \code{colour}
}
\examples{
library(ggplot2)

example_plot <-
ggplot(diamonds, aes(x = cut, y = price, colour = cut)) +
  stat_summary(fun = "mean", geom = "bar", fill = "white")

# plain palette
example_plot + scale_colour_uq_d(palette = "discrete")

# with transparency
example_plot + scale_colour_uq_d(palette = "discrete",
                                 alpha = 0.6)

# reverse palette order
example_plot + scale_colour_uq_d(palette = "discrete",
                                 reverse = TRUE)
}
