% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uq.R
\name{scale_fill_uq_c}
\alias{scale_fill_uq_c}
\title{Continuous ggplot2 palette for \code{fill}}
\usage{
scale_fill_uq_c(
  palette = c("main", "secondary", "bw", "discrete", "continuous", "neutral"),
  alpha = 1,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Palette name (one of \code{main}, \code{secondary}, \code{bw}, \code{discrete}, \code{continuous}, or \code{neutral}). Default: \code{main}}

\item{alpha}{Colour transparency. Default: 1 (no transparency)}

\item{reverse}{If \code{TRUE}, reverse the colour palette order. Default: \code{FALSE}}

\item{...}{Additional arguments to pass to \code{ggplot2::discrete_scale()}}
}
\value{
A continuous ggplot2 palette for \code{fill}
}
\description{
Continuous ggplot2 palette for \code{fill}
}
\examples{
library(ggplot2)

example_plot <-
ggplot(faithfuld,
       aes(waiting, eruptions, fill = density)) +
       geom_tile()

# plain palette
example_plot + scale_fill_uq_c(palette = "continuous")

# with transparency
example_plot + scale_fill_uq_c(palette = "continuous",
                               alpha = 0.6)

# reverse palette order
example_plot + scale_fill_uq_c(palette = "continuous",
                               reverse = TRUE)
}
